[config] logger_config own file
per title
