+++ config/src/config/logger_config.rs
@@ -1,0 +1,22 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(default)]
+pub struct LoggerConfig {
+    // Use async logging
+    pub is_async: bool,
+    // chan_size of slog async drain for node logging.
+    pub chan_size: usize,
+}
+
+impl Default for LoggerConfig {
+    fn default() -> LoggerConfig {
+        LoggerConfig {
+            is_async: true,
+            chan_size: 256,
+        }
+    }
+}
